server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-socketio==5.11.1
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.9.3
jinja2==3.1.3